BPM is now a float value for greater accuracy with external sync

+++ Sequencer/EatsSwingUtils.h
@@ -14,6 +14,6 @@
 + (NSArray *) swingArray;
 
 // Returns the amount of time in nanoseconds a note should be pushed back
-+ (uint64_t) calculateSwingNsForPosition:(uint)position type:(int)swingType amount:(int)swingAmount bpm:(uint)bpm qnPerMeasure:(uint)qnPerMeasure minQuantization:(uint)minQuantization;
++ (uint64_t) calculateSwingNsForPosition:(uint)position type:(int)swingType amount:(int)swingAmount bpm:(float)bpm qnPerMeasure:(uint)qnPerMeasure minQuantization:(uint)minQuantization;
 
 @end
